LearningPoint/learningPoint/requirements.txt
Django==3.2
djangorestframework==3.14.0
six==1.10.0
